// packages/datadog-api-client-v1/models/HostMetrics.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Host Metrics collected.
*/
export declare class HostMetrics {
    /**
    * The percent of CPU used (everything but idle).
    */
    'cpu'?: number;
    /**
    * The percent of CPU spent waiting on the IO (not reported for all platforms).
    */
    'iowait'?: number;
    /**
    * The system load over the last 15 minutes.
    */
    'load'?: number;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): HostMetrics;
    static serialize(data: HostMetrics): {
        [key: string]: any;
    };
    constructor();
}